+++ ga/vrpdfd/utils/cpp_utils.pyi
@@ -13,9 +13,9 @@ __all__ = (
 
 
 def set_customers(
-    low: Sequence[float],
-    high: Sequence[float],
-    w: Sequence[float],
+    low: Sequence[int],
+    high: Sequence[int],
+    w: Sequence[int],
     x: Sequence[float],
     y: Sequence[float],
 ) -> None: ...
@@ -24,22 +24,22 @@ def set_customers(
 def paths_from_flow(
     truck_paths_count: int,
     drone_paths_count: Sequence[int],
-    flows: Sequence[Sequence[float]],
+    flows: Sequence[Sequence[int]],
     neighbors: Sequence[Set[int]],
 ) -> Tuple[
-    List[Dict[int, float]],
-    List[List[Dict[int, float]]],
+    List[Dict[int, int]],
+    List[List[Dict[int, int]]],
 ]: ...
 
 
 def paths_from_flow_chained(
     truck_paths: Sequence[AbstractSet[int]],
     drone_paths: Sequence[Sequence[AbstractSet[int]]],
-    truck_capacity: float,
-    drone_capacity: float,
+    truck_capacity: int,
+    drone_capacity: int,
 ) -> Tuple[
-    List[Dict[int, float]],
-    List[List[Dict[int, float]]],
+    List[Dict[int, int]],
+    List[List[Dict[int, int]]],
 ]: ...
